+++ frontend/src/styles/baseComponents.js
@@ -6,7 +6,10 @@ const baseTable = styled.table`
     border-collapse: collapse;
     border-collapse: collapse;
     padding: 10px 10px;
-    border: 1px solid black;
+    /* border: 1px solid black; */
+    border: 1px solid #e5e5e5;
+    border-radius: 4px 4px 0px 0px;
+    -moz-border-radius: 4px 4px 0px 0px;
     text-align: left;
     vertical-align: middle;
     word-wrap: break-word;
@@ -14,28 +17,36 @@ const baseTable = styled.table`
     td:first-of-type {
         vertical-align: top;
     }
+    /* td:not(:first-child) {
+        border-top: none;
+    } */
     td.test-result-undefined {
         background: #eee;
     }
 
-    td {
-        border-collapse: collapse;
-        padding: 10px 10px;
-        border: 1px solid black;
-        text-align: left;
-        vertical-align: middle;
-        word-wrap: break-word;
-        background: #ffffff;
-    }
-
+    td,
     th {
         border-collapse: collapse;
         padding: 10px 10px;
-        border: 1px solid black;
+        /* border: 1px solid black; */
+        /* border-top: 1px solid #e5e5e5; */
         text-align: left;
         vertical-align: middle;
         word-wrap: break-word;
-        background: #ddd;
+    }
+
+    td {
+        background: #ffffff;
+        /* border-top: 1px solid #e5e5e5; */
+    }
+
+    th {
+        background: #f0f0f0;
+        /* border-top: none; */
+    }
+
+    tr {
+        border-top: 1px solid #e5e5e5;
     }
 `;
